Modify CSS and JS files
Created home.js to handle creating and loading the "Home" tab when the page loads. The elements added through the JS here are just tests at the moment.
Modified index.js to import the function from home.js.
Modified styles.css to add flex to some of the divs so that the home page looks how I want it to. The styles are just tests to differentiate some of the elements.