Modify HTML and JS
Created contact.js and menu.js which will hold the logic to add those pages to the DOM. They have some placeholder elements for now.
Modified index.js to import the functions for loading the other tabs. Added event listeners to each of the nav buttons to load the correct page when clicked.
Modified template.html to add IDs to each of the nav buttons.